Some type  conversions.

+++ card.cpp
@@ -6,7 +6,7 @@ void Card::setType(CardType inputType){
     type = inputType;
 }
 
-void Card::setPower(unsigned short inputPower){
+void Card::setPower(float inputPower){
     if(inputPower < 0){
         inputPower = 0;
     }
@@ -17,6 +17,6 @@ CardType Card::getType()const{
     return type;
 }
 
-unsigned short Card::getPower()const{
+float Card::getPower()const{
     return power;
 }
